backend/django/requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
django-bootstrap5==24.2
pyotp==2.9.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
